+++ lib/src/page_transition.dart
@@ -10,9 +10,15 @@ class PageTransition<T> extends PageRouteBuilder<T> {
   /// Child for your next page
   final Widget child;
 
+  /// Child for your next page
+  final Widget childCurrent;
+
   /// Transition types
   ///  fade,rightToLeft,leftToRight, upToDown,downToUp,scale,rotate,size,rightToLeftWithFade,leftToRightWithFade
   final PageTransitionType type;
+
+  /// Transition types
+  final PageTransitionType reverseType;
 
   /// Curves for transitions
   final Curve curve;
@@ -22,6 +28,9 @@ class PageTransition<T> extends PageRouteBuilder<T> {
 
   /// Durationf for your transition default is 300 ms
   final Duration duration;
+
+  /// Duration for your pop transition default is 300 ms
+  final Duration reverseDuration;
 
   /// Context for inheret theme
   final BuildContext ctx;
@@ -34,11 +43,14 @@ class PageTransition<T> extends PageRouteBuilder<T> {
     Key key,
     @required this.child,
     @required this.type,
+    this.reverseType,
+    this.childCurrent,
     this.ctx,
     this.inheritTheme = false,
     this.curve = Curves.linear,
     this.alignment,
     this.duration = const Duration(milliseconds: 300),
+    this.reverseDuration = const Duration(milliseconds: 300),
     RouteSettings settings,
   })  : assert(inheritTheme ? ctx != null : true,
             "'ctx' cannot be null when 'inheritTheme' is true, set ctx: context"),
@@ -53,6 +65,7 @@ class PageTransition<T> extends PageRouteBuilder<T> {
                 : child;
           },
           transitionDuration: duration,
+          reverseTransitionDuration: reverseDuration,
           settings: settings,
           opaque: false,
           transitionsBuilder: (BuildContext context,
@@ -67,115 +80,44 @@ class PageTransition<T> extends PageRouteBuilder<T> {
               /// PageTransitionType.rightToLeft which is the give us right to left transition
               case PageTransitionType.rightToLeft:
                 return SlideTransition(
-                  transformHitTests: false,
-                  position: new Tween<Offset>(
-                    begin: const Offset(1.0, 0.0),
+                  position: Tween<Offset>(
+                    begin: const Offset(1, 0),
                     end: Offset.zero,
-                  ).animate(
-                    CurvedAnimation(
-                      parent: animation,
-                      curve: curve,
-                    ),
-                  ),
-                  child: new SlideTransition(
-                    position: new Tween<Offset>(
-                      begin: Offset.zero,
-                      end: const Offset(-1.0, 0.0),
-                    ).animate(
-                      CurvedAnimation(
-                        parent: secondaryAnimation,
-                        curve: curve,
-                      ),
-                    ),
-                    child: child,
-                  ),
+                  ).animate(animation),
+                  child: child,
                 );
                 break;
 
               /// PageTransitionType.leftToRight which is the give us left to right transition
-
               case PageTransitionType.leftToRight:
                 return SlideTransition(
-                  transformHitTests: false,
                   position: Tween<Offset>(
-                    begin: const Offset(-1.0, 0.0),
+                    begin: const Offset(-1, 0),
                     end: Offset.zero,
-                  ).animate(
-                    CurvedAnimation(
-                      parent: animation,
-                      curve: curve,
-                    ),
-                  ),
-                  child: new SlideTransition(
-                    position: new Tween<Offset>(
-                      begin: Offset.zero,
-                      end: const Offset(1.0, 0.0),
-                    ).animate(
-                      CurvedAnimation(
-                        parent: secondaryAnimation,
-                        curve: curve,
-                      ),
-                    ),
-                    child: child,
-                  ),
+                  ).animate(animation),
+                  child: child,
                 );
                 break;
 
-              /// PageTransitionType.upToDown which is the give us up to down transition
-
+              /// PageTransitionType.topToBottom which is the give us up to down transition
               case PageTransitionType.topToBottom:
                 return SlideTransition(
-                  transformHitTests: false,
                   position: Tween<Offset>(
-                    begin: const Offset(0.0, -1.0),
+                    begin: const Offset(0, -1),
                     end: Offset.zero,
-                  ).animate(
-                    CurvedAnimation(
-                      parent: animation,
-                      curve: curve,
-                    ),
-                  ),
-                  child: new SlideTransition(
-                    position: new Tween<Offset>(
-                      begin: Offset.zero,
-                      end: const Offset(0.0, 1.0),
-                    ).animate(
-                      CurvedAnimation(
-                        parent: secondaryAnimation,
-                        curve: curve,
-                      ),
-                    ),
-                    child: child,
-                  ),
+                  ).animate(animation),
+                  child: child,
                 );
                 break;
 
               /// PageTransitionType.downToUp which is the give us down to up transition
-
               case PageTransitionType.bottomToTop:
                 return SlideTransition(
-                  transformHitTests: false,
                   position: Tween<Offset>(
-                    begin: const Offset(0.0, 1.0),
+                    begin: const Offset(0, 1),
                     end: Offset.zero,
-                  ).animate(
-                    CurvedAnimation(
-                      parent: animation,
-                      curve: curve,
-                    ),
-                  ),
-                  child: new SlideTransition(
-                    position: new Tween<Offset>(
-                      begin: Offset.zero,
-                      end: const Offset(0.0, -1.0),
-                    ).animate(
-                      CurvedAnimation(
-                        parent: secondaryAnimation,
-                        curve: curve,
-                      ),
-                    ),
-                    child: child,
-                  ),
+                  ).animate(animation),
+                  child: child,
                 );
                 break;
 
@@ -199,10 +141,10 @@ class PageTransition<T> extends PageRouteBuilder<T> {
               /// PageTransitionType.rotate which is the rotate functionality for transition you can also use alignment for this transition
 
               case PageTransitionType.rotate:
-                return new RotationTransition(
+                return RotationTransition(
                   alignment: alignment,
                   turns: animation,
-                  child: new ScaleTransition(
+                  child: ScaleTransition(
                     alignment: alignment,
                     scale: animation,
                     child: FadeTransition(
@@ -240,14 +182,9 @@ class PageTransition<T> extends PageRouteBuilder<T> {
                     opacity: animation,
                     child: SlideTransition(
                       position: Tween<Offset>(
-                        begin: Offset.zero,
-                        end: const Offset(-1.0, 0.0),
-                      ).animate(
-                        CurvedAnimation(
-                          parent: secondaryAnimation,
-                          curve: curve,
-                        ),
-                      ),
+                        begin: const Offset(1, 0),
+                        end: Offset.zero,
+                      ).animate(animation),
                       child: child,
                     ),
                   ),
@@ -271,17 +208,89 @@ class PageTransition<T> extends PageRouteBuilder<T> {
                     opacity: animation,
                     child: SlideTransition(
                       position: Tween<Offset>(
-                        begin: Offset.zero,
-                        end: const Offset(1.0, 0.0),
+                        begin: const Offset(-1, 0),
+                        end: Offset.zero,
+                      ).animate(animation),
+                      child: child,
+                    ),
+                  ),
+                );
+                break;
+
+              case PageTransitionType.rightToLeftJoined:
+                assert(childCurrent != null, """
+                When using type "rightToLeftJoined" you need argument: 'childCurrent'
+
+                example:
+                  child: MyPage(),
+                  childCurrent: this
+
+                """);
+                return Stack(
+                  children: <Widget>[
+                    SlideTransition(
+                      position: Tween<Offset>(
+                        begin: const Offset(0.0, 0.0),
+                        end: const Offset(-1.0, 0.0),
                       ).animate(
                         CurvedAnimation(
-                          parent: secondaryAnimation,
+                          parent: animation,
+                          curve: curve,
+                        ),
+                      ),
+                      child: childCurrent,
+                    ),
+                    SlideTransition(
+                      position: Tween<Offset>(
+                        begin: const Offset(1.0, 0.0),
+                        end: Offset.zero,
+                      ).animate(
+                        CurvedAnimation(
+                          parent: animation,
+                          curve: curve,
+                        ),
+                      ),
+                      child: child,
+                    )
+                  ],
+                );
+                break;
+
+              case PageTransitionType.leftToRightJoined:
+                assert(childCurrent != null, """
+                When using type "leftToRightJoined" you need argument: 'childCurrent'
+                example:
+                  child: MyPage(),
+                  childCurrent: this
+
+                """);
+                return Stack(
+                  children: <Widget>[
+                    SlideTransition(
+                      position: Tween<Offset>(
+                        begin: const Offset(-1.0, 0.0),
+                        end: const Offset(0.0, 0.0),
+                      ).animate(
+                        CurvedAnimation(
+                          parent: animation,
                           curve: curve,
                         ),
                       ),
                       child: child,
                     ),
-                  ),
+                    SlideTransition(
+                      position: Tween<Offset>(
+                        begin: const Offset(0.0, 0.0),
+                        end: const Offset(1.0, 0.0),
+                      ).animate(
+                        CurvedAnimation(
+                          parent: animation,
+                          curve: curve,
+                        ),
+                      ),
+                      child: childCurrent,
+                    )
+                  ],
                 );
                 break;