+++ CHANGELOG.md
@@ -84,3 +84,11 @@
 
 ## [1.1.7+2] - Update
 - update readme
+## [1.1.7+3] - Added reverseTransitionDuration
+- Now you can set pop duration.
+## [1.1.7+4] - Added new Transition methods
+- leftToRightJoined and rightToLeftJoined 
+## [1.1.7+5] - Remove null explicitly initialize
+-  Remove null explicitly initialize
+## [1.1.7+6] - Fix issues
+-  Fix incorrect transition and hero animation issue.